requirements.txt



streamlit
openai
oracledb
pandas
matplotlib
plotly
faiss-cpu
tqdm
python-dotenv


